# webapp/APIVoo/impl/swagger-spec.json
{"openapi":"3.0.0","paths":{"/reservas":{"post":{"operationId":"ReservaController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservaInsertDTO"}}}},"responses":{"200":{"description":"Reserva criada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservaDTO"}}}}},"tags":["Reservas"]}},"/reservas/{id}":{"get":{"operationId":"ReservaController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Reserva retornada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservaDTO"}}}}},"tags":["Reservas"]},"delete":{"operationId":"ReservaController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Reserva deletada"}},"tags":["Reservas"]}},"/voos":{"post":{"operationId":"VooController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VooInsertDTO"}}}},"responses":{"200":{"description":"Voo criado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VooDTO"}}}}},"tags":["Voos"]},"get":{"operationId":"VooController_findAll","parameters":[],"responses":{"200":{"description":"Voos retornados","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VooDTO"}}}}}},"tags":["Voos"]}},"/voos/{id}":{"get":{"operationId":"VooController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Voo retornado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VooDTO"}}}}},"tags":["Voos"]},"put":{"operationId":"VooController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voo"}}}},"responses":{"200":{"description":"Voo atualizado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VooDTO"}}}}},"tags":["Voos"]},"delete":{"operationId":"VooController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Voos"]}},"/voos/{id}/assentos":{"get":{"operationId":"VooController_findAssentosByBooId","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Voo e seus assentos retornado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VooAssentosDTO"}}}}},"tags":["Voos"]}},"/assentos":{"post":{"operationId":"AssentoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssentoInsertDTO"}}}},"responses":{"200":{"description":"Assento criado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssentoDTO"}}}}},"tags":["Assentos"]},"get":{"operationId":"AssentoController_findAll","parameters":[],"responses":{"200":{"description":"Assentos retornados","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AssentoDTO"}}}}}},"tags":["Assentos"]}},"/assentos/{id}":{"get":{"operationId":"AssentoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Assento retornado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssentoDTO"}}}}},"tags":["Assentos"]},"put":{"operationId":"AssentoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssentoUpdateDTO"}}}},"responses":{"200":{"description":"Assento atualizado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssentoDTO"}}}}},"tags":["Assentos"]},"delete":{"operationId":"AssentoController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Assento deletado"}},"tags":["Assentos"]}}},"info":{"title":"AC2","description":"AC2 Programação Distribuída","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ReservaInsertDTO":{"type":"object","properties":{"nome":{"type":"string","example":"Thiago"},"sobrenome":{"type":"string","example":"Sanches"},"rg":{"type":"string","example":"12345678-9"},"vooId":{"type":"number","example":1},"assentoId":{"type":"number","example":1}},"required":["nome","sobrenome","rg","vooId","assentoId"]},"ReservaDTO":{"type":"object","properties":{"id":{"type":"number","example":1},"nome":{"type":"string","example":"Thiago Sanches"}},"required":["id","nome"]},"VooInsertDTO":{"type":"object","properties":{"origem":{"type":"string","example":"Origem do voo"},"destino":{"type":"string","example":"Destino do voo"},"data":{"format":"date-time","type":"string","example":"2023-10-10"}},"required":["origem","destino","data"]},"VooDTO":{"type":"object","properties":{"id":{"type":"number","example":1},"origem":{"type":"string","example":"Origem do voo"},"destino":{"type":"string","example":"Destino do voo"},"data":{"format":"date-time","type":"string","example":"2023-10-10"}},"required":["id","origem","destino","data"]},"AssentoDTO":{"type":"object","properties":{"fileira":{"type":"string","example":1},"numero":{"type":"number","example":1},"reservado":{"type":"boolean","example":true}},"required":["fileira","numero","reservado"]},"VooAssentosDTO":{"type":"object","properties":{"id":{"type":"number","example":1},"origem":{"type":"string","example":"Origem do voo"},"destino":{"type":"string","example":"Destino do voo"},"data":{"format":"date-time","type":"string","example":"2023-10-10"},"assentos":{"type":"array","items":{"$ref":"#/components/schemas/AssentoDTO"}}},"required":["id","origem","destino","data","assentos"]},"Voo":{"type":"object","properties":{}},"AssentoInsertDTO":{"type":"object","properties":{"fileira":{"type":"string","example":"A"},"numero":{"type":"number","example":1},"vooId":{"type":"number","example":1}},"required":["fileira","numero","vooId"]},"AssentoUpdateDTO":{"type":"object","properties":{"fileira":{"type":"string","example":"A"},"numero":{"type":"number","example":1}},"required":["fileira","numero"]}}}}